V0.2
Prototype is completed. It performs basic "Match3" mechanics with the addition of 4 match that creates explosion effect at the left most side or the down most side of the link and the addition of empty grid mechanic.
A "Level Editor" have been added. Levels can be editted and created using it.

- Empty grids have been added. ("IsHole"). These grids does not have any grid object, or background. All functions works according to that.
 e.g. Newly created gems skips empty grids, or matching link functions ignore empty grids.